requirement.txt
jupyterlab==4.5.1
numpy==2.3.3
matplotlib==3.10.7
pyvisa==1.15.0
ipykernel==7.1.0
pyserial==3.5
pandas==2.3.3